Enable -Wextra (without -Wsign-compare)

+++ src/crypto/equihash.tcc
@@ -25,11 +25,11 @@ bool IsProbablyDuplicate(std::shared_ptr<eh_trunc> indices, size_t lenIndices)
 {
     assert(lenIndices <= MAX_INDICES);
     bool checked_index[MAX_INDICES] = {false};
-    int count_checked = 0;
-    for (int z = 0; z < lenIndices; z++) {
+    size_t count_checked = 0;
+    for (size_t z = 0; z < lenIndices; z++) {
         // Skip over indices we have already paired
         if (!checked_index[z]) {
-            for (int y = z+1; y < lenIndices; y++) {
+            for (size_t y = z+1; y < lenIndices; y++) {
                 if (!checked_index[y] && indices.get()[z] == indices.get()[y]) {
                     // Pair found
                     checked_index[y] = true;
